refactor: split `QuestionsController` into partial classes for better organization
- Divided `QuestionsController` into multiple partial classes (`Level`, `Location`, `Organisation`, `When`) to improve maintainability.
- Updated related unit tests to reflect the changes.
- Adjusted namespaces and imports as needed.
